# openapi.yaml
openapi: 3.0.3
info:
  title: Sunset Beach Resort & Spa — Admin/Booking API
  description: >
    Contract reverse-engineered from the current Next.js route handlers under
    `app/api/**/route.ts`. It documents exactly what each route accepts and
    returns today (including quirks such as Prisma `Decimal`/`Date` JSON
    serialization), so it can serve as a migration contract for a backend
    rewrite in another language. No behavior was changed to produce this spec.
  version: "1.0.0"
  license:
    name: Proprietary
  x-codegen-notes: >
    A property marked `x-sensitive: true` (guest email/phone, payment notes) is redacted as
    "[REDACTED]" in its generated model's toString() by the customized pojo.mustache template
    under openapi-templates/ - see scripts/generate-api.sh, which passes --template-dir. This is
    the one and only place PII redaction needs to be declared; it survives regeneration
    automatically. Do not hand-edit a generated model's toString() to redact a field instead -
    mark the field here and regenerate.
servers:
  - url: /api
    description: Relative to the deployed Next.js app origin.
...
security:
  - bearerAuth: []
paths:
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [Menu]
      summary: List all menu items
      description: Requires any authenticated staff session (ADMIN/MANAGER/CASHIER/WAITER).
      responses:
        "200":
          description: All menu items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createMenuItem
      tags: [Menu]
      summary: Create a menu item
      description: Requires role `MANAGER` or above.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Menu item created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
  /menu/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getMenuItem
      tags: [Menu]
      summary: Get a menu item by id
      description: Requires any authenticated staff session.
      responses:
        "200":
          description: The menu item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Menu item not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Menu item not found
    patch:
      operationId: updateMenuItem
      tags: [Menu]
      summary: Update a menu item
      description: Requires role `MANAGER` or above. Full replacement of all `MenuItemInput` fields — no partial update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "200":
          description: Updated menu item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
        "404":
          description: Menu item not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Menu item not found
    delete:
      operationId: deleteMenuItem
      tags: [Menu]
      summary: Delete a menu item
      description: Requires role `MANAGER` or above.
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
        "404":
          description: Menu item not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Menu item not found
        "409":
          description: This menu item has existing order lines and can't be deleted (FK constraint).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: This menu item has existing order lines and can't be deleted.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT returned by `POST /auth/login`, sent as `Authorization: Bearer
        <token>`. Contains `sub` (user id), `email`, `role`, standard `iat`/`exp`
        (default TTL 7 days, configurable via `JWT_TTL_DAYS`). Every route in
        this file except `POST /bookings` and the unauthenticated `/auth/*`
        endpoints (`login`, `logout`) requires a valid token, returning 401 if
        missing/invalid/expired, and 403 where the route additionally requires
        role `ADMIN`.
  responses:
    Unauthorized:
      description: No valid JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            error: Unauthorized
...
    ForbiddenManagerPlus:
      description: Token is valid but lacks the required role (`MANAGER` or above).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            error: Forbidden
...
  schemas:
...
    MenuDepartment:
      type: string
      enum: [KITCHEN, BAR]
      description: >
        Where a MenuItem's kitchen/bar ticket prints - independent of `category`, which is a
        free-text display grouping for the menu itself (e.g. "Mains", "Cocktails") and has no
        effect on print routing.
...
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
      description: 'Generic error shape used by all hand-written `{ error: "..." }` responses.'
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          description: Output of Zod's `error.flatten()`.
          properties:
            formErrors:
              type: array
              items:
                type: string
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
      example:
        error:
          formErrors: []
          fieldErrors:
            guestEmail: ["Invalid email"]
    OkTrue:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
...
    MenuItem:
      type: object
      required: [id, name, description, category, department, price, isAvailable, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        department:
          $ref: "#/components/schemas/MenuDepartment"
        price:
          type: string
          description: Decimal(10,2) rendered as a string, e.g. `"250.00"` — same convention as `Room.basePrice`.
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MenuItemInput:
      type: object
      description: >
        Full replacement on PATCH — no partial update (same convention as `RoomInput`).
        `department` defaults to `KITCHEN` (most menu items are food) - existing rows migrated
        the same way; staff reassign drinks to `BAR` explicitly.
      required: [name, description, category, price]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 2000
        category:
          type: string
          minLength: 1
          maxLength: 60
        department:
          $ref: "#/components/schemas/MenuDepartment"
          default: KITCHEN
        price:
          type: number
          minimum: 0
          exclusiveMinimum: true
        isAvailable:
          type: boolean
          default: true
...