# src/main/java/com/sunsetbeach/api/MenuApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sunsetbeach.api;

import com.sunsetbeach.model.ErrorMessage;
import com.sunsetbeach.model.MenuItem;
import com.sunsetbeach.model.MenuItemInput;
import com.sunsetbeach.model.OkTrue;
import com.sunsetbeach.model.ValidationError;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
public interface MenuApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /menu : Create a menu item
     * Requires role &#x60;MANAGER&#x60; or above.
     *
     * @param menuItemInput  (required)
     * @return Menu item created. (status code 201)
     *         or Body failed validation. (status code 400)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/menu",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<MenuItem> createMenuItem(
         @Valid @RequestBody MenuItemInput menuItemInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"isAvailable\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : \"price\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"category\" : \"category\", \"department\" : \"KITCHEN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"formErrors\" : [ ], \"fieldErrors\" : { \"guestEmail\" : [ \"Invalid email\" ] } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /menu/{id} : Delete a menu item
     * Requires role &#x60;MANAGER&#x60; or above.
     *
     * @param id  (required)
     * @return Deleted. (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     *         or Menu item not found. (status code 404)
     *         or This menu item has existing order lines and can&#39;t be deleted (FK constraint). (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/menu/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<OkTrue> deleteMenuItem(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"ok\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /menu/{id} : Get a menu item by id
     * Requires any authenticated staff session.
     *
     * @param id  (required)
     * @return The menu item. (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Menu item not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menu/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<MenuItem> getMenuItem(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"isAvailable\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : \"price\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"category\" : \"category\", \"department\" : \"KITCHEN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /menu : List all menu items
     * Requires any authenticated staff session (ADMIN/MANAGER/CASHIER/WAITER).
     *
     * @return All menu items. (status code 200)
     *         or No valid JWT. (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menu",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<MenuItem>> listMenuItems(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"isAvailable\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : \"price\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"category\" : \"category\", \"department\" : \"KITCHEN\" }, { \"isAvailable\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : \"price\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"category\" : \"category\", \"department\" : \"KITCHEN\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /menu/{id} : Update a menu item
     * Requires role &#x60;MANAGER&#x60; or above. Full replacement of all &#x60;MenuItemInput&#x60; fields — no partial update.
     *
     * @param id  (required)
     * @param menuItemInput  (required)
     * @return Updated menu item. (status code 200)
     *         or Body failed validation. (status code 400)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     *         or Menu item not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/menu/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<MenuItem> updateMenuItem(
         @PathVariable("id") String id,
         @Valid @RequestBody MenuItemInput menuItemInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"isAvailable\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : \"price\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"category\" : \"category\", \"department\" : \"KITCHEN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"formErrors\" : [ ], \"fieldErrors\" : { \"guestEmail\" : [ \"Invalid email\" ] } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}